minor css mod

+++ Corso/M4/D7-8/script/backoffice.js
@@ -140,19 +140,21 @@ const printProduct = (allProducts) => {
 }
 
 async function deleteProduct(id) {
-  try {
-    const response = await fetch(`${apiUrl}product/${id}`, {
-      method: "DELETE",
-      headers: {
-        "Authorization": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NGU0ZjUzZmRmZmI4YjAwMTQ0MTNkMzUiLCJpYXQiOjE2OTI3MjY1OTIsImV4cCI6MTY5MzkzNjE5Mn0.Rv-6TAE7YE7A5tkUA8TnwiK8eQ6Gt70j2AuLUMsJdVs"
+  if (confirm('Sei sicuro di voler eliminare questo prodotto?')) {
+    try {
+      const response = await fetch(`${apiUrl}product/${id}`, {
+        method: "DELETE",
+        headers: {
+          "Authorization": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NGU0ZjUzZmRmZmI4YjAwMTQ0MTNkMzUiLCJpYXQiOjE2OTI3MjY1OTIsImV4cCI6MTY5MzkzNjE5Mn0.Rv-6TAE7YE7A5tkUA8TnwiK8eQ6Gt70j2AuLUMsJdVs"
+        }
+      })
+  
+      if (response.ok) {
+        window.location.href = 'backoffice.html'
       }
-    })
-
-    if (response.ok) {
-      window.location.href = 'backoffice.html'
+    } catch (error) {
+      console.log('Errore durante eliminazione prodotto: ', error);
+      alert('Si è verificato un errore durante eliminazione.')
     }
-  } catch (error) {
-    console.log('Errore durante eliminazione prodotto: ', error);
-    alert('Si è verificato un errore durante eliminazione.')
   }
 }
